Continued development for adding calories
updated main to support navigation in dart for switching between pages in the flutter app. I also added a calorie_tracking_page.dart file which is the view for the new page for adding calories and a smaple button to go between a "home page" and the calorie tracking page.